requirements.txt
accelerate==0.18.0
boto3==1.28.62
botocore==1.31.62
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.7
clip==1.0
cramjam==2.7.0
cycler==0.11.0
dpcpp-cpp-rt==2023.2.0
faiss-cpu==1.7.0
fastparquet==0.8.1
filelock==3.12.0
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
intel-cmplr-lib-rt==2023.2.0
intel-cmplr-lic-rt==2023.2.0
intel-opencl-rt==2023.2.0
intel-openmp==2023.2.0
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.5.3
mkl==2023.2.0
mkl-fft==1.3.1
mkl-service==2.4.0
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.1
pandas==1.1.5
Pillow==9.0.1
protobuf==4.24.2
psutil==5.9.5
pyarrow==12.0.1
pybase64==1.2.3
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-transformers==1.0.0
pytrec-eval==0.5
pytz==2023.3.post1
PyYAML==6.0
regex==2022.10.31
requests==2.29.0
s3transfer==0.7.0
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0.post9
tbb==2021.10.0
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.22.2
typing_extensions==4.7.1
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
